refactor(frontend): move Optimizer predicates into opt/ headers
optimizeChunkOnce() carried ~300 lines of read-only predicates inline:
opcode traits (isJump/isTerminator/isStoreOp/canWriteVariables and the two
operator-name tables) plus the folding predicates (popEntry/identityEligible/
foldPureMath/addConstant/emitPush).

They now live under frontend/ir/opt/analysis/. OpTraits keeps the four hot
one-liners inline in the header and only the two name tables out of line;
FoldPredicates keeps the trivial variant accessors inline and moves the rest
to a .cpp. TrackedValue and StackEntry move to opt/OptContext.h so both can
see them.

Pure move: no call site changed its behaviour, and the bytecode dump over the
52-script corpus stays byte-identical.
