+++ src/LOICollectionA/frontend/ir/Optimizer.cpp
@@ -1,265 +1,20 @@
-#include <cmath>
-#include <limits>
 #include <vector>
 #include <variant>
 #include <unordered_map>
 #include <unordered_set>
 
-#include "LOICollectionA/utils/core/MathUtils.h"
+#include "LOICollectionA/frontend/DiagnosticEngine.h"
 
 #include "LOICollectionA/frontend/ir/VM.h"
+
+#include "LOICollectionA/frontend/ir/opt/OptContext.h"
+#include "LOICollectionA/frontend/ir/opt/analysis/FoldPredicates.h"
+#include "LOICollectionA/frontend/ir/opt/analysis/OpTraits.h"
 
 #include "LOICollectionA/frontend/ir/Optimizer.h"
 
 namespace LOICollection::frontend::ir {
-namespace {
-    struct TrackedValue {
-        ValueNode::ValueType value;
-        int producer = -1;
-        bool removable = false;
-    };
-
-    using StackEntry = std::variant<TrackedValue, std::monostate>;
-
-    bool isJump(OpCode op) {
-        return op == OpCode::JMP || op == OpCode::JMP_IF_FALSE || op == OpCode::JMP_IF_TRUE;
-    }
-
-    bool isTerminator(OpCode op) {
-        return op == OpCode::JMP || op == OpCode::RETURN || op == OpCode::HALT;
-    }
-
-    bool isKnown(const StackEntry& entry) {
-        return std::holds_alternative<TrackedValue>(entry);
-    }
-
-    bool isScalarValue(const ValueNode::ValueType& value) {
-        return !std::holds_alternative<ArrayRef>(value) &&
-            !std::holds_alternative<ObjectRef>(value) &&
-            !std::holds_alternative<FunctionRefPtr>(value);
-    }
-
-    const TrackedValue& knownValue(const StackEntry& entry) {
-        return std::get<TrackedValue>(entry);
-    }
-
-    StackEntry popEntry(std::vector<StackEntry>& stack) {
-        if (stack.size() > 1) {
-            StackEntry top = std::move(stack.back());
-            stack.pop_back();
-            return top;
-        }
-
-        return std::monostate{};
-    }
-
-    std::string arithmeticOpName(OpCode op) {
-        switch (op) {
-            case OpCode::ADD: return "+";
-            case OpCode::SUB: return "-";
-            case OpCode::MUL: return "*";
-            case OpCode::DIV: return "/";
-            case OpCode::MOD: return "%";
-            case OpCode::POW: return "^";
-            default: return "";
-        }
-    }
-
-    std::string comparisonOpName(OpCode op) {
-        switch (op) {
-            case OpCode::CMP_EQ: return "==";
-            case OpCode::CMP_NE: return "!=";
-            case OpCode::CMP_GT: return ">";
-            case OpCode::CMP_LT: return "<";
-            case OpCode::CMP_GE: return ">=";
-            case OpCode::CMP_LE: return "<=";
-            default: return "";
-        }
-    }
-
-    // Opcodes that may execute script code (function bodies, lambdas, constructors) or
-    // create object fields; any of these can rebind what a later LOAD_VAR resolves to.
-    bool canWriteVariables(OpCode op) {
-        switch (op) {
-            case OpCode::CALL:
-            case OpCode::CALL_MACRO:
-            case OpCode::CALL_METHOD:
-            case OpCode::CALL_METHOD_VIRTUAL:
-            case OpCode::CALL_FUNC:
-            case OpCode::CALL_LAMBDA:
-            case OpCode::CALL_SUPER_CTOR:
-            case OpCode::NEW:
-            case OpCode::STORE_FIELD:
-                return true;
-            default:
-                return false;
-        }
-    }
-
-    bool isStoreOp(OpCode op) {
-        return op == OpCode::STORE_VAR || op == OpCode::STORE_SLOT;
-    }
-
-    // `kept op identity == kept`, applicable when the kept operand is a known number of a
-    // compatible kind. Divided by type-promotion rules: `x / 1` and `x ^ 1` always produce
-    // floats, and float `x + 0.0` flips -0.0, so those combinations stay untouched.
-    bool identityEligible(OpCode op, const StackEntry& kept, const StackEntry& identity) {
-        if (!isKnown(kept) || !isKnown(identity) || !knownValue(identity).removable)
-            return false;
-
-        const auto& keptValue = knownValue(kept).value;
-        const auto& identityValue = knownValue(identity).value;
-
-        bool keptIsInt = std::holds_alternative<int>(keptValue);
-        bool keptIsFloat = std::holds_alternative<float>(keptValue);
-        bool identityIntZero = std::holds_alternative<int>(identityValue) && std::get<int>(identityValue) == 0;
-        bool identityIntOne = std::holds_alternative<int>(identityValue) && std::get<int>(identityValue) == 1;
-        bool identityFloatOne = std::holds_alternative<float>(identityValue) && std::get<float>(identityValue) == 1.0f;
-
-        switch (op) {
-            case OpCode::ADD: // x + 0 / 0 + x
-                return keptIsInt && identityIntZero;
-            case OpCode::SUB: // x - 0
-                return keptIsInt && identityIntZero;
-            case OpCode::MUL: // x * 1 / 1 * x
-                return (keptIsInt && identityIntOne) || (keptIsFloat && (identityIntOne || identityFloatOne));
-            case OpCode::DIV: // x / 1
-                return keptIsFloat && (identityIntOne || identityFloatOne);
-            case OpCode::POW: // x ^ 1
-                return keptIsFloat && (identityIntOne || identityFloatOne);
-            default:
-                return false;
-        }
-    }
-
-    // ValueType holds non-comparable alternatives, so scalar equality is checked per kind.
-    // Floats compare bitwise-sign aware: `+0.0 == -0.0` under IEEE754, but they are
-    // observably different values (`to_string` and `1/x` disagree), so pool dedup must
-    // never conflate them.
-    bool sameScalar(const ValueNode::ValueType& a, const ValueNode::ValueType& b) {
-        if (a.index() != b.index() || !isScalarValue(a))
-            return false;
-
-        switch (a.index()) {
-            case 0: return std::get<int>(a) == std::get<int>(b);
-            case 1: {
-                const float fa = std::get<float>(a);
-                const float fb = std::get<float>(b);
-                return fa == fb && std::signbit(fa) == std::signbit(fb);
-            }
-            case 2: return std::get<std::string>(a) == std::get<std::string>(b);
-            case 3: return std::get<bool>(a) == std::get<bool>(b);
-            default: return true;
-        }
-    }
-
-    bool foldPureMath(
-        const std::string& name,
-        const std::vector<ValueNode::ValueType>& args,
-        ValueNode::ValueType& out
-    ) {
-        auto isInt = [](const ValueNode::ValueType& v) { return std::holds_alternative<int>(v); };
-        auto isFloat = [](const ValueNode::ValueType& v) { return std::holds_alternative<float>(v); };
-
-        if (name == "math::abs" && args.size() == 1) {
-            if (isInt(args[0])) {
-                int v = std::get<int>(args[0]);
-                if (v == std::numeric_limits<int>::min())
-                    return false;
-
-                out = std::abs(v);
-                return true;
-            }
-
-            if (isFloat(args[0])) {
-                out = std::abs(std::get<float>(args[0]));
-                return true;
-            }
-
-            return false;
-        }
-
-        if ((name == "math::min" || name == "math::max") && args.size() == 2) {
-            if (isInt(args[0]) && isInt(args[1])) {
-                out = name == "math::min"
-                    ? std::min(std::get<int>(args[0]), std::get<int>(args[1]))
-                    : std::max(std::get<int>(args[0]), std::get<int>(args[1]));
-                return true;
-            }
-
-            if (isFloat(args[0]) && isFloat(args[1])) {
-                out = name == "math::min"
-                    ? std::min(std::get<float>(args[0]), std::get<float>(args[1]))
-                    : std::max(std::get<float>(args[0]), std::get<float>(args[1]));
-                return true;
-            }
-
-            return false;
-        }
-
-        if ((name == "math::sqrt" || name == "math::log" || name == "math::sin" ||
-             name == "math::cos") && args.size() == 1) {
-            if (!isInt(args[0]) && !isFloat(args[0]))
-                return false;
-
-            double v = isInt(args[0]) ? std::get<int>(args[0]) : std::get<float>(args[0]);
-            out = static_cast<float>(
-                name == "math::sqrt" ? std::sqrt(v) :
-                name == "math::log" ? std::log(v) :
-                name == "math::sin" ? std::sin(v) : std::cos(v)
-            );
-            return true;
-        }
-
-        if (name == "math::pow" && args.size() == 2) {
-            if (isInt(args[0]) && isInt(args[1])) {
-                out = static_cast<float>(
-                    MathUtils::pow(std::get<int>(args[0]), std::get<int>(args[1]))
-                );
-                return true;
-            }
-
-            if (isFloat(args[0]) && isFloat(args[1])) {
-                out = static_cast<float>(std::pow(std::get<float>(args[0]), std::get<float>(args[1])));
-                return true;
-            }
-
-            return false;
-        }
-
-        return false;
-    }
-
-    int addConstant(BytecodeChunk& chunk, const ValueNode::ValueType& value) {
-        if (isScalarValue(value)) {
-            for (size_t i = 0; i < chunk.constants.size(); ++i) {
-                if (sameScalar(chunk.constants[i], value))
-                    return static_cast<int>(i);
-            }
-        }
-
-        chunk.constants.push_back(value);
-        return static_cast<int>(chunk.constants.size() - 1);
-    }
-
-    void emitPush(
-        BytecodeChunk& chunk, std::vector<Instruction>& out, const ValueNode::ValueType& value,
-        const SourceLocation& loc = {}
-    ) {
-        OpCode op = OpCode::PUSH_INT;
-
-        switch (value.index()) {
-            case 0: op = OpCode::PUSH_INT; break;
-            case 1: op = OpCode::PUSH_FLOAT; break;
-            case 2: op = OpCode::PUSH_STR; break;
-            case 3: op = OpCode::PUSH_BOOL; break;
-            case 7: op = OpCode::PUSH_NONE; break;
-            default: op = OpCode::PUSH_INT; break;
-        }
-
-        out.push_back({ op, addConstant(chunk, value), loc });
-    }
-}
+    using namespace opt;
 
     Optimizer::Stats Optimizer::optimize(BytecodeChunk& chunk) {
         Stats total;
